refactor: Separate Seat type definition into its own file
Move the Seat type definition from seats.tsx to a new file seat.ts to improve code organization and clarity. Update imports accordingly.
